+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HumHub Changelog
 -------------------
 - Fix #5851: Admin User Settings cannot saved 
 - Fix #5856: Fix SMTPS Config Migration
+- Fix #5854: Don't validate user language on approve
 
 
 1.12.1 (August 15, 2022)